Create getAll and getPaginated methods for the Family Member feature, including their implementation in the repository, resource, and factory.

+++ database/migrations/2025_05_29_145855_create_head_of_families_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
                 ->onDelete('cascade');
 
             $table->string('profile_picture')->nullable();
-            $table->integer('identity_number')->unique();
+            $table->bigInteger('identity_number')->unique();
             $table->enum('gender', ['male', 'female']);
             $table->date('date_of_birth')->nullable();
             $table->string('phone_number')->nullable();
